# Conflicts:
#	src/cli/cmdline_route.c
#	src/graph/geometry.h
#	src/routing.cpp
#	src/routing.h

+++ src/make_pruned_gl.cpp
@@ -1,4 +1,4 @@
-#include "routing.h"
+#include "interface/Client.h"
 
 #include <fstream>
 #include <chrono>
@@ -10,7 +10,7 @@ int main(int argc, const char *argv[]) {
     std::string graph_file(argv[1]);
     std::string output_file(argv[2]);
     std::ofstream output(output_file);
-    float epsilon = std::stof(argv[3]);
+    double epsilon = std::stof(argv[3]);
     Client client;
     client.read_graph_file(graph_file, epsilon, false);
